[FE] refactor: views 파일 구조화 #115
